Refactor: split a component
