+++ src/components/common/Modal.tsx
@@ -30,20 +30,27 @@ const Modal: React.FC<ModalProps> = ({ children }) => {
         !isOpenMainModal &&
         createPortal(
           <div>
-            <div onClick={() => closeModal()} className="fixed top-0 left-0 w-full h-full bg-black opacity-50 z-20"></div>
-            <div className="bg-white fixed top-1/2 left-1/2 px-[5.5rem] py-12 flex flex-col justify-center items-start rounded-2xl translate-x-[-50%] translate-y-[-50%] z-30">{children}</div>
+            <div onClick={() => closeModal()} className="fixed top-0 left-0 w-full h-full bg-[#00000060] z-20">
+              <div onClick={e => e.stopPropagation()} className="bg-white fixed top-1/2 left-1/2 px-[5.5rem] py-12 flex flex-col justify-center items-start rounded-2xl translate-x-[-50%] translate-y-[-50%] z-30">
+                {children}
+              </div>
+            </div>
           </div>,
-          document.body,
+          document.getElementById('modal-portal') as HTMLDivElement,
+          // document.body,
         )}
 
       {isOpenMainModal &&
         createPortal(
           <div>
-            <div onClick={() => mainCloseModal()} className="fixed top-0 left-0 w-full h-full bg-black opacity-50 z-20"></div>
-            <div className="bg-white fixed top-1/2 left-1/2 px-[5.5rem] py-12 flex flex-col justify-center items-start rounded-2xl translate-x-[-50%] translate-y-[-50%] z-30">{children}</div>
+            <div onClick={() => mainCloseModal()} className="fixed top-0 left-0 w-full h-full bg-[#00000060] z-20">
+              <div onClick={e => e.stopPropagation()} className="bg-white fixed top-1/2 left-1/2 px-[5.5rem] py-12 flex flex-col justify-center items-start rounded-2xl translate-x-[-50%] translate-y-[-50%] z-30">
+                {children}
+              </div>
+            </div>
           </div>,
-          document.body,
-          // document.getElementById('modal-portal') as HTMLDivElement,
+          document.getElementById('modal-portal') as HTMLDivElement,
+          // document.body,
         )}
     </>
   );
