--- src/components/ideaPostPage/useIdeaPost.hook.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function useIdeaPost() {
-  return <div>useIdeaPost</div>;
-}
-
-export default useIdeaPost;
